requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.5
django-celery-results==2.2.0
djangorestframework==3.12.4
flake8==3.9.2
greenlet==1.1.0
imagecorruptions==1.1.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
kombu==5.1.0
matplotlib==3.4.2
mccabe==0.6.1
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
Shapely==1.7.1
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
tifffile==2021.7.2
vine==5.0.0
wcwidth==0.2.5
